Fix bug introduced by a5378b9bf27e2e3a6812aea5163243e417c03f23
If the python files and schema were modifed the CMake files that contain
the generated targets would not be recreated. Now the CMake code
explicitly maintains a list of files that the generation of the target
files depends on so that if any of those files change regeneration
is forced.