Implement support for declaring dependency handlers external from the svcb folder by using the new `add_cmake_dependency_handler()` macro.
Using this macro will cause all benchmarks in the current directory
and below to be aware of how to handle the dependency that the handler
declared in the file passed to `add_cmake_dependency_handler()`
knows how to handle.

This mechanism works by dynamically loading the file passed to
`add_cmake_dependency_handler()` as python code and executing it.