Change what the default for the `exhaustive_counter_examples` is. Now if `correct` is false then it is implicitly true if counter examples are specified and is implicitly false if none are provided.
The motivation behind this change is that the default is not very
beneficial because the idea behind having counter examples is to
try to prevent a verifier from pretending it found a bug by checking
the counter example it gives. Previously the default was non exhaustive
so this information couldn't actually be used to our advantage.

Technically this is change deserves a schema version bump however
that's quite a bit of work and I don't think anyone was relying
on the behaviour of this attribute so it's probably early enough
that we can make this change without adversely effecting users.

Another change here is that the implicit behaviour is made explicit
when creating a benchmark object. The result of doing this is that
generated augmented spec files will explicitly provide this property.