Remove using `-fprofile-dir=` for compilation during a profile build.
This isn't very helpful because the dependencies (e.g. libGSL) don't
get this flag so the coverage information (`*.gcda` files) end up
in the build tree.

Instead it seems that setting the `GCOV_PREFIX` environment variable is
a much simpler to solve this problem. The `*.gcda` files end up
where there would normally in the build tree but prefixed with
whatever `GCOV_PREFIX` is.