Introduce a wrapper for the `stdlib.h` header so we can interpose calls to `abort()` and make sure `__gcov_flush()` is called first.
This tries to ensure that coverage information will be flushed to disk
if we hit an `abort()` call. This won't work for libraries linked to
dynamically or that the build system didn't instrument (e.g. if glibc
calls `abort()` internally) but for our purposes this will probably work
most of the time.